- moving com.deltek.integration.maconomy.domain classes to com.deltek.integration.maconomy.domain.internal, where they should not be exposed and remain internal - copying interfaces from com.maconomy.web.client to com.deltek.integration.maconomy.domain - copying classes from com.maconomy.web.client.util to com.deltek.integration.maconomy.util
The next step would be to begin integrating the interfaces from
com.deltek.integration.maconomy.domain with the corresponding objects
from com.deltek.integration.maconomy.domain.internal. We need to decide
if we would like to have POJO classes that are used for data exchanges
with web service to implement those interfaces directly, or if there
should be classes in between that have references to POJO objects and
which implement the interfaces.