Starting to base the implementation of classes in com.deltek.integration.maconomy.domain.internal on interfaces located in com.deltek.integration.maconomy.domain. Beginning with RecordImpl which is now not using template type.
Card, Table correspond to PaneState, we might consider introducing
CardPaneState and TablePaneState or just keep PaneState if that does not
introduce any advantage. Card and Pane share most of the attributes,
that is why I was thinking about having one entity.

Introducing string and integer Maconomy values, just a quick attempt to
fill in the gap, their interface and implementation is open to change.

Removing LinkProvider in favour of HasLinks, as they are the same
concepts.

Tests in MaconomyPSORestContextTest use Mockito to model changes in the
data object, which is not exposing possibility to change its state.